+++ scripts/utils/add-staff-user.js
@@ -40,6 +40,13 @@ userService
     console.log(opts.username + ' staff status:', staffStatus);
     console.log('done');
   })
+  // wait 5 seconds to allow for asynchronous `event-listeners` to finish
+  // https://github.com/troupe/gitter-webapp/issues/580#issuecomment-147445395
+  // https://gitlab.com/gitterHQ/webapp/merge_requests/1605#note_222861592
+  .then(() => {
+    console.log(`Waiting 5 seconds to allow for the asynchronous \`event-listeners\` to finish...`);
+    return new Promise(resolve => setTimeout(resolve, 5000));
+  })
   .finally(function() {
     shutdown.shutdownGracefully();
   });