+++ scripts/utils/auto-lurk-room.js
@@ -11,6 +11,8 @@ var Promise = require('bluebird');
 var shutdown = require('shutdown');
 var es = require('event-stream');
 
+require('../../server/event-listeners').install();
+
 var opts = require('yargs')
   .option('room', {
     alias: 'r',
@@ -104,7 +106,13 @@ function handleMultipleRooms() {
 }
 
 run()
-  .delay(1000)
+  // wait 5 seconds to allow for asynchronous `event-listeners` to finish
+  // https://github.com/troupe/gitter-webapp/issues/580#issuecomment-147445395
+  // https://gitlab.com/gitterHQ/webapp/merge_requests/1605#note_222861592
+  .then(() => {
+    console.log(`Waiting 5 seconds to allow for the asynchronous \`event-listeners\` to finish...`);
+    return new Promise(resolve => setTimeout(resolve, 5000));
+  })
   .then(function() {
     shutdown.shutdownGracefully();
   })