+++ scripts/utils/delete-messages-from-user.js
@@ -14,6 +14,8 @@ const chatService = require('gitter-web-chats');
 const troupeService = require('gitter-web-rooms/lib/troupe-service');
 const chatsForUserSearch = require('gitter-web-elasticsearch/lib/chats-for-user-search');
 
+require('../../server/event-listeners').install();
+
 const opts = require('yargs')
   .option('username', {
     alias: 'u',
@@ -109,6 +111,13 @@ const clearMessages = async () => {
 };
 
 clearMessages()
+  // wait 5 seconds to allow for asynchronous `event-listeners` to finish
+  // https://github.com/troupe/gitter-webapp/issues/580#issuecomment-147445395
+  // https://gitlab.com/gitterHQ/webapp/merge_requests/1605#note_222861592
+  .then(() => {
+    console.log(`Waiting 5 seconds to allow for the asynchronous \`event-listeners\` to finish...`);
+    return new Promise(resolve => setTimeout(resolve, 5000));
+  })
   .then(function() {
     shutdown.shutdownGracefully();
   })