+++ scripts/utils/unread-remove-impossible.js
@@ -12,6 +12,8 @@ var _ = require('lodash');
 /** THIS SCRIPT WILL TRASH OUR REDIS ENVIRONMENT */
 process.exit();
 
+require('../../server/event-listeners').install();
+
 function getKeys() {
   var redisClient = redis.getClient();
 
@@ -86,11 +88,16 @@ function markAllWeirdRoomsAsRead(hash) {
 getKeys()
   .then(convertToHash)
   .then(markAllWeirdRoomsAsRead)
-  .delay(1000)
+  // wait 5 seconds to allow for asynchronous `event-listeners` to finish
+  // https://github.com/troupe/gitter-webapp/issues/580#issuecomment-147445395
+  // https://gitlab.com/gitterHQ/webapp/merge_requests/1605#note_222861592
+  .then(() => {
+    console.log(`Waiting 5 seconds to allow for the asynchronous \`event-listeners\` to finish...`);
+    return new Promise(resolve => setTimeout(resolve, 5000));
+  })
   .catch(function(err) {
     console.error(err.stack);
   })
-  .delay(10000)
   .finally(function() {
     shutdown.shutdownGracefully();
   });