+++ scripts/utils/update-room-tags.js
@@ -9,6 +9,8 @@ var Promise = require('bluebird');
 
 var roomTagger = require('../../server/utils/room-tagger');
 
+require('../../server/event-listeners').install();
+
 // @const
 var BATCH_SIZE = 20;
 var QUERY_LIMIT = 0;
@@ -159,11 +161,22 @@ function run(rooms) {
 
   if (CALLED_RUN % (BATCH_SIZE * 5) === 0) logProgress();
 
-  return attachRepoInfoToRooms(rooms)
-    .then(filterFailedRooms)
-    .then(tagRooms)
-    .then(saveRooms)
-    .catch(function(err) {
-      if (err.statusCode !== 404) console.error(err.stack);
-    });
+  return (
+    attachRepoInfoToRooms(rooms)
+      .then(filterFailedRooms)
+      .then(tagRooms)
+      .then(saveRooms)
+      // wait 5 seconds to allow for asynchronous `event-listeners` to finish
+      // https://github.com/troupe/gitter-webapp/issues/580#issuecomment-147445395
+      // https://gitlab.com/gitterHQ/webapp/merge_requests/1605#note_222861592
+      .then(() => {
+        console.log(
+          `Waiting 5 seconds to allow for the asynchronous \`event-listeners\` to finish...`
+        );
+        return new Promise(resolve => setTimeout(resolve, 5000));
+      })
+      .catch(function(err) {
+        if (err.statusCode !== 404) console.error(err.stack);
+      })
+  );
 }