+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 21.40.0 - 2021-08-02
+
+ - Disallow invites from hellbanned users, https://gitlab.com/gitterHQ/webapp/-/merge_requests/2237
+
+Developer facing:
+
+ - Be consistent about waiting for the asyncronous event-listeners to finish in util scripts, https://gitlab.com/gitterHQ/webapp/-/merge_requests/2234
+
+
 # 21.39.0 - 2021-07-16
 
  - Update to `matrix-appservice-bridge@2.7.0`, https://gitlab.com/gitterHQ/webapp/-/merge_requests/2235